+++ setup-pipwn.sh
@@ -92,7 +92,9 @@ uninstall_pipwn() {
   
   rm -f /etc/nginx/sites-enabled/pipwn
   rm -f /etc/nginx/sites-available/pipwn
-  sudo mv /etc/nginx/sites-available/default.disabled /etc/nginx/sites-enabled/default 2>/dev/null || true
+  if [ -f /etc/nginx/sites-available/default.disabled ]; then
+    sudo mv /etc/nginx/sites-available/default.disabled /etc/nginx/sites-enabled/default 2>/dev/null || true
+  fi
   systemctl reload nginx 2>/dev/null || true
   
   echo ""
